requirements.txt
# Backend
fastapi
uvicorn[standard]
langchain
langchain-google-genai
pypdf
faiss-cpu
python-dotenv

# Frontend
streamlit
requests